pyproject.toml
[tool.poetry]
name = "storybot"
version = "0.1.0"
description = "Telegram bot for Story-Point tracking"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "bot" },
    { include = "core" },
    { include = "db" },
    { include = "integrations" },
]

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.7"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
aiofiles = "^23.2.1"
loguru = "^0.7.2"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
bot = "bot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88